+++ src/Php/PhpType.php
@@ -13,4 +13,14 @@ class PhpType {
         $this->name = $name;
         $this->alias = $alias;
     }
+
+    public function equals(PhpType $other): bool {
+        if ($this->namespace !== $other->namespace) {
+            return false;
+        }
+        if (str_replace('[]', '', $this->name) !== str_replace('[]', '', $other->name)) {
+            return false;
+        }
+        return true;
+    }
 }