+++ src/Relation.php
@@ -33,13 +33,25 @@ class Relation {
     }
 
     public function getRelations(): array {
-        $classNames = array_map(function($x){ return $x->class->getClassType()->name; }, $this->namespace->getEntries());
-        $arrows = array_filter($this->namespace->getArrows(), function($x) use ($classNames) {
-            return in_array(str_replace('[]', '', $x->to), $classNames);
-        });
-        $relation_expressions = array_map(function($x){
-            return $x->toString();
-        }, $arrows);
+        $entities = $this->namespace->getEntries();
+        // TODO クラスの同一性をクラス名のみで判定している。本当はnamespaceも含めて判定する必要がある。
+//        $classNames = array_map(function($x){
+//            return $x->class->getClassType()->name;
+//        }, $entities);
+//        // 解析対象に含まれているクラスで絞り込む
+//        $arrows = array_filter($this->namespace->getArrows(), function($x) use ($classNames) {
+//            return in_array(str_replace('[]', '', $x->to->name), $classNames);
+//        });
+        $relation_expressions = array_map(function($x) use ($entities){
+            foreach ($entities as $e) {
+                //if ($e->class->getClassType()->name == str_replace('[]', '', $x->to->name)) {
+                if ($e->class->getClassType()->equals($x->to)) {
+                    return $x->toString($e->class);
+                }
+            }
+            return null;
+        }, $this->namespace->getArrows());
+        $relation_expressions = array_filter($relation_expressions);
         sort($relation_expressions);
         return $relation_expressions;
     }