+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ### Added
 
  * Added dependency arrows. The class has a dependency on the return type of the method with no public arguments.
+ * Changed the output class name to include the package name.
+
+### Bug fix
+ * fix bug that failed to parse NullableType method parameter.
+ * fix bug that wrong namespace found when search use statements.
 
 ## v0.0.1 (2021-11-11)